pyproject.toml
[project]
name = "lattice"
version = "0.1.0"
description = "Graph-powered code intelligence - hybrid RAG with graph and vector search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tree-sitter-language-pack>=0.7.1",
    "openai>=1.0.0",
    "qdrant-client>=1.12.0",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tiktoken>=0.8.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.28.0",
    "rapidfuzz>=3.5.0",
    "asyncpg>=0.29.0",
    "claude-agent-sdk>=0.1.0",
    "langchain-text-splitters>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
lattice = "lattice.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lattice"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
